+++ modules/spring-web/src/test/kotlin/com/esprito/spring/web/completion/java/UastMockMvcUrlReferenceContributorTest.kt
@@ -13,7 +13,8 @@ class UastMockMvcUrlReferenceContributorTest : EspritoJavaLightTestCase() {
         get() = arrayOf(
             TestLibrary.springWeb_6_0_7,
             TestLibrary.springTest_6_0_7,
-            TestLibrary.springBootAutoConfigure_3_1_1
+            TestLibrary.springBootAutoConfigure_3_1_1,
+            TestLibrary.springReactiveWeb_3_1_1
         )
 
     fun testPut() {
@@ -161,4 +162,81 @@ class UastMockMvcUrlReferenceContributorTest : EspritoJavaLightTestCase() {
         val classFqn = resolvedElement?.containingClass?.qualifiedName
         TestCase.assertEquals("OrderController#getOrders", "$classFqn#$memberName")
     }
+
+    fun testSimpleRouterFunction() {
+        myFixture.copyFileToProject("WebClients.java")
+        myFixture.configureByText(
+            "WebClientTest.java", """
+            import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+            
+            class WebClientTest {
+                void justForTest() {
+                    MockMvcRequestBuilders.get("/us<caret>ers");
+                }
+            }
+        """.trimIndent()
+        )
+
+        val ref = file.findReferenceAt(myFixture.caretOffset) as? EspritoControllerMethodReference
+        assertNotNull(ref)
+        val multiResolve = ref!!.multiResolve(true)
+        assertEquals(1, multiResolve.size)
+        val resolvedElement = (multiResolve[0].element) as? PsiMember
+        assertNotNull(resolvedElement)
+        val memberName = resolvedElement?.name
+        val classFqn = resolvedElement?.containingClass?.qualifiedName
+        TestCase.assertEquals("WebClients#oneRoute", "$classFqn#$memberName")
+    }
+
+
+    fun testDiffPostRouterFunction() {
+        myFixture.copyFileToProject("WebClients.java")
+        myFixture.configureByText(
+            "WebClientTest.java", """
+            import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+            
+            class WebClientTest {
+                void justForTest() {
+                    MockMvcRequestBuilders.post("/users/cus<caret>tomers");
+                }
+            }
+        """.trimIndent()
+        )
+
+        val ref = file.findReferenceAt(myFixture.caretOffset) as? EspritoControllerMethodReference
+        assertNotNull(ref)
+        val multiResolve = ref!!.multiResolve(true)
+        assertEquals(1, multiResolve.size)
+        val resolvedElement = (multiResolve[0].element) as? PsiMember
+        assertNotNull(resolvedElement)
+        val memberName = resolvedElement?.name
+        val classFqn = resolvedElement?.containingClass?.qualifiedName
+        TestCase.assertEquals("WebClients#differentFunction", "$classFqn#$memberName")
+    }
+
+    fun testDiffDeleteRouterFunction() {
+        myFixture.copyFileToProject("WebClients.java")
+        myFixture.configureByText(
+            "WebClientTest.java", """
+            import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+            
+            class WebClientTest {
+                void justForTest() {
+                    MockMvcRequestBuilders.delete("/users/{use<caret>rId}");
+                }
+            }
+        """.trimIndent()
+        )
+
+        val ref = file.findReferenceAt(myFixture.caretOffset) as? EspritoControllerMethodReference
+        assertNotNull(ref)
+        val multiResolve = ref!!.multiResolve(true)
+        assertEquals(1, multiResolve.size)
+        val resolvedElement = (multiResolve[0].element) as? PsiMember
+        assertNotNull(resolvedElement)
+        val memberName = resolvedElement?.name
+        val classFqn = resolvedElement?.containingClass?.qualifiedName
+        TestCase.assertEquals("WebClients#differentFunction", "$classFqn#$memberName")
+    }
+
 }